requirements.txt
langchain
langchain-core
langchain-openai==0.2.9
langchain-google-genai
langchain-community
langchain-text-splitters
langchain-experimental
openai
python-dotenv
langchain-huggingface
transformers
tokenizers
torch
sentence-transformers
scikit-learn
streamlit
pydantic 
grandalf
pypdf2
